+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * Fixed zlib encoder on web
 * Added `CodecsRecognizer` with `ArchiveFormat` to recognize compressed file types with their header bytes
 * Fixed `XZEncoder` (without compression) for large files
+* Fixed SHA-256 checks for XZDecoder with `verify: true`
+* Increased SHA-256 speed for XZDecoder and XZEncoder
 * Made CRC32 faster on IO (non-js-web platforms)
 * Slightly reduces amount of RAM for lzma decoder and zlib encoder
 * Added `ProgressOutputStream` to monitor progress during unpack